requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
PyMySQL
Werkzeug
cryptography
itsdangerous
flask_migrate
Flask-JWT-Extended
requests
redis